rename folder to match package name
